Refactor service package
