+++ src/main/java/lits/jp/hotel/management/models/StaffMember.java
@@ -15,13 +15,13 @@ public class StaffMember {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int staffMemberId;
 
-    @Column(name = "first_name")
+    @Column(name = "first_name", nullable = false)
     private String firstName;
 
-    @Column(name = "last_name")
+    @Column(name = "last_name", nullable = false)
     private String lastName;
 
-    @Column(name = "password")
+    @Column(name = "password", nullable = false)
     private String password;
 
     @Column(name = "role")
