+++ src/test/java/lits/jp/hotel/management/services/GuestServiceImplTest.java
@@ -1,9 +1,8 @@
-package lits.jp.hotel.management.service;
+package lits.jp.hotel.management.services;
 
 import lits.jp.hotel.management.repository.GuestsRepository;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
